i18n small refactor
